Refactor stubbed Home page components into separate files
