+++ FrontEnd/src/components/Help.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Help extends React.Component {
+  render() {
+    return <p>Help</p>;
+  }
+}
+
+export default Help;
