+++ FrontEnd/src/components/Stats.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Stats extends React.Component {
+  render() {
+    return <p>Stats</p>;
+  }
+}
+
+export default Stats;
